+++ docs/lib/Components/index.js
@@ -50,8 +50,8 @@ class Components extends React.Component {
           to: '/components/breadcrumbs/'
         },
         {
-          name: 'Tags',
-          to: '/components/tags/'
+          name: 'Badge',
+          to: '/components/badge/'
         },
         {
           name: 'Card',
@@ -122,12 +122,12 @@ class Components extends React.Component {
   }
   render() {
     return (
-      <Container fluid className="content">
+      <Container className="content">
         <Row>
           <Col md={{ size: 3, push: 9 }}>
             <div className="docs-sidebar mb-3">
               <h5>Components</h5>
-              <Nav>
+              <Nav className="flex-column">
                 {this.state.navItems.map((item, i) => {
                   return <ComponentLink key={i} item={item} />;
                 })}