+++ docs/lib/NotFound/index.js
@@ -7,8 +7,8 @@ export default () => {
   return (
     <div>
       <Helmet title="404 Page Not Found" />
-      <section className="jumbotron text-xs-center mb-3">
-        <Container fluid>
+      <section className="jumbotron text-center mb-3">
+        <Container>
           <Row>
             <Col>
               <p className="lead">