+++ src/__tests__/Alert.spec.js
@@ -12,9 +12,9 @@ describe('Alert', () => {
     const alert = shallow(<Alert>Yo!</Alert>);
     expect(alert.prop('transitionName')).toEqual({
       appear: 'fade',
-      appearActive: 'in',
+      appearActive: 'show',
       enter: 'fade',
-      enterActive: 'in',
+      enterActive: 'show',
       leave: 'fade',
       leaveActive: 'out'
     });