+++ src/__tests__/Collapse.spec.js
@@ -33,9 +33,14 @@ describe('Collapse', () => {
     expect(wrapper.hasClass('collapse')).toEqual(true);
   });
 
-  it('should render with class "in" when isOpen is true', () => {
+  it('should render with class "navbar"', () => {
+    const wrapper = shallow(<Collapse navbar />);
+    expect(wrapper.hasClass('navbar')).toEqual(true);
+  });
+
+  it('should render with class "show" when isOpen is true', () => {
     const wrapper = shallow(<Collapse isOpen />);
-    expect(wrapper.hasClass('in')).toEqual(true);
+    expect(wrapper.hasClass('show')).toEqual(true);
   });
 
   it('should set height to null when isOpen is true', () => {